Finished C# dynamic client proxy.
Cleanup whitespace and unused imports.

+++ Java/Services/src/common/InfoService.java
@@ -5,6 +5,6 @@ import ToyORB.ToyORBService;
 @SuppressWarnings({"MethodParameterNamingConvention", "InstanceMethodNamingConvention", "InterfaceNamingConvention"})
 public interface InfoService extends ToyORBService
 {
-    int getTemperature(String city);
+    float getTemperature(String city);
     String getRoadInfo(int roadId);
 }
